Extract feature list to its own viewmodel
